Add assignNum function and call it in newText
The function assignNum will check if the incoming string is an operator,
if so it will assign it to calculate.operator. If not then the function
check if num1 is null, if so it assigns the string to it. Finally if not
then it checks if num2 is null, if so assign the string to num2.