Rework calculator logic to work as expected
Target is now a global let variable so it can be changed and read across
the code. Anytime a number is set in calculate.numX, an op btn is
clicked, or an operation is performed calculate.result is added and set
to true. When a button is clicked the first check is if result in
calculate exists, if so delete result from calculate and remove the
highlight from target. If the current target (previous op btn clicked)
is not equals then set display.textContent to "".

During an operation when num1 and num2 aren't null now the operation
function such as add will set num1 to operation result and num2 to null.
Then in the operation function the display will be set to the result.
Lastly the operator is set to null and if the target.id is equal then
set num1 to null, as equal was pressed and a new calculation is
happening.

Also when a num is set the display.textContent is set to itself just to
make sure it is set to that; and comments were added.